requirements.txt
requests
lxml
django
gunicorn
python-dateutil
django-ratelimit
django-registration
django-bower
django_nvd3
django-bootstrap-pagination
psycopg2
